+++ magda-sleuther-broken-link/src/index.ts
@@ -6,7 +6,14 @@ import commonYargs from "@magda/sleuther-framework/dist/commonYargs";
 
 const ID = "sleuther-broken-link";
 
-const argv = commonYargs(ID, 6111, "http://localhost:6111");
+const argv = commonYargs(ID, 6111, "http://localhost:6111", argv =>
+    argv.option("externalRetries", {
+        describe:
+            "Number of times to retry external links when checking whether they're broken",
+        type: "number",
+        default: 3
+    })
+);
 
 function sleuthBrokenLinks() {
     return sleuther({
@@ -17,7 +24,7 @@ function sleuthBrokenLinks() {
         async: true,
         writeAspectDefs: [brokenLinkAspectDef, datasetQualityAspectDef],
         onRecordFound: (record, registry) =>
-            onRecordFound(record, registry, argv.retries)
+            onRecordFound(record, registry, argv.externalRetries)
     });
 }